ref(layouts): Rename "layout" to "layouts"
- Renamed the "src/layout" -> "src/layouts"
  - This prevent name confusion with "components/layout" directory
